Completed the microservices project of QUiz application with eureka loadbalancer and fiegn

+++ Question-Service/src/main/java/com/growandshine/Question_Service/Controller/QuestionController.java
@@ -43,7 +43,7 @@ public class QuestionController {
     }
 
     @GetMapping("/generate")
-    public ResponseEntity<List<Integer>> generateQuestions(@RequestParam String category,@RequestParam long numberofquestions){
+    public ResponseEntity<List<Long>> generateQuestions(@RequestParam String category,@RequestParam long numberofquestions){
         return questionService.generateQuestionsForQuiz(category,numberofquestions);
     }
